Fix: 로또 결과 도메인 수익률 타입 변경
Double 최대 범위 초과로 인한 값 변경 이슈

+++ src/main/java/service/LottoResultService.java
@@ -18,7 +18,7 @@ public class LottoResultService {
         }
 
         BigDecimal totalProfit = lottoResult.createTotalProfit(lottoResult);
-        Double profitRate = lottoResult.createProfitRate(totalProfit ,userLotto.getAmount());
+        BigDecimal profitRate = lottoResult.createProfitRate(totalProfit ,userLotto.getAmount());
 
         lottoResult.setTotalProfitAndProfitRate(totalProfit, profitRate);
